Add validation directory
